backend/requirements.txt
# Core
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
pydantic-settings==2.1.0

# Database
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
alembic==1.12.1

# Authentication
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6

# Whisper
openai-whisper==20231117
openai==1.3.0

# Groq (for LLM action item extraction)
groq==0.4.2

# Audio processing
ffmpeg-python==0.2.0

# External APIs
requests==2.31.0

# Utilities
python-dotenv==1.0.0

